esdoc chandes and minor refactor -add required esdoc comments for source -rename widthHook.js to useWidth.js -replace LanguageSelector.js from src/components/Header to src/tools -rename src/components/header to src/components/Header -update type validations in several files
