+++ src/components/Content/Content.js
@@ -2,6 +2,10 @@ import React from 'react'
 import ResultArea from "./WorkSpace/ResultArea";
 // import MainContainer from "./Carusel/Container";
 
+/**
+ * main path of th app
+ * @returns {jsx}
+ */
 const Content = ()=> (<div>
     <ResultArea/>
 </div>);
