+++ src/components/Footer/Footer.js
@@ -2,6 +2,12 @@ import React from "react";
 import {withNamespaces} from "react-i18next";
 import PropTypes from 'prop-types';
 
+/**
+ * @ignore
+ * @param t
+ * @returns {*}
+ * @constructor
+ */
 const Footer = ({t}) =>(<div>{t("footer")}</div>);
 
 Footer.propTypes = {
